+++ src/pages/Tenants.tsx
@@ -1,9 +1,12 @@
+import React, { useEffect, useState } from "react";
 import DatePicker from "../components/DatePicker";
 import Searchbar from "../components/Searchbar";
 import ApartmentSelector from "../components/ApartmentSelector";
 import TenantTable from "../components/TenantTable";
 import Button from "../components/Button";
+import Pagination from "../components/Pagination";
 import { ButtonVarient } from "../types/ButtonTypes";
+import { User } from "../types/TableTypes";
 
 interface TenantsProps {
   selectedView: string;
@@ -19,6 +22,22 @@ const Tenants: React.FC<TenantsProps> = ({ selectedView }) => {
     "Last Signed In",
     "Actions",
   ];
+
+  const [tenants, setTenants] = useState<User[]>([]);
+  const [currentPage, setCurrentPage] = useState(1);
+  const [rowsPerPage, setRowsPerPage] = useState(25);
+
+  useEffect(() => {
+    fetch("assets/data.json")
+      .then((response) => response.json())
+      .then((data) => setTenants(data))
+      .catch((error) => console.error("error fetching data:", error));
+  }, []);
+
+  const totalTenants = tenants.length;
+  const startIndex = (currentPage - 1) * rowsPerPage;
+  const endIndex = Math.min(currentPage * rowsPerPage, totalTenants);
+  const currentTenants = tenants.slice(startIndex, endIndex);
 
   return (
     <>
@@ -43,7 +62,21 @@ const Tenants: React.FC<TenantsProps> = ({ selectedView }) => {
             </div>
             <div className="my-1 h-px bg-gray-200"></div>
           </div>
-          <TenantTable headers={headers} hasIcon={true} hasButtons={true} />
+
+          <TenantTable
+            headers={headers}
+            hasIcon={true}
+            hasButtons={true}
+            tenants={currentTenants}
+          />
+
+          <Pagination
+            data={tenants}
+            currentPage={currentPage}
+            rowsPerPage={rowsPerPage}
+            setCurrentPage={setCurrentPage}
+            setRowsPerPage={setRowsPerPage}
+          />
         </>
       ) : (
         <p className="flex items-center">No data available</p>
