+++ src/types/TableTypes.ts
@@ -14,4 +14,5 @@ export interface TenantTableProps {
   headers: string[];
   hasIcon?: boolean;
   hasButtons?: boolean;
+  tenants: User[];
 }
